src/dataMongoDB/requirement.txt
pymongo[srv]==3.11
python-dotenv
certifi